BugID(s): VOIP-64473
Problem Description:
Don’t display the error (or return error return code) message if the phone is doing –aa and receives a register: this is by design!

Fix Description:
Added a test case to check to make sure -aa is doing what we think it should.  I.E. when the client recieves the unexpected register (whilst paused) it automatically replies.  The test still passes.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@23336 85c8ee13-dc6b-4ec3-b49e-93a6737f2991