BugID(s): VOIP-64473
Problem Description:
If script stops early due to 'Q' being pressed this is FAIL, not PASS (change SIPp return code).

Fix Description:
Added a check to make sure quitting is zero before setting it to 1 because the number of calls has reached the number the user wanted to run.  abortcall now reports the call as failed.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@23495 85c8ee13-dc6b-4ec3-b49e-93a6737f2991