BugID(s): VOIP-64473
Problem Description:
SIPp to listen only for traffic FROM a specific address. This will allow it to ignore unrelated traffic (like REGISTER requests from other phones). Must also allow multiple instances running at the same time.

Fix Description:
Made a change for the case when -mc is enabled, and SIPp is running as a client.  In this case the socket can be connected to the given remote address so all other traffic is ignored.  Made a change to the sipp do connect function, to let connect be called on UDP sockets when only one call is being run.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@23595 85c8ee13-dc6b-4ec3-b49e-93a6737f2991