BugID(s): VOIP-64473
Problem Description:
Added error checking when you call parse offset for a media_port_offset in a pcap play.  Note that the when you call parse offset from the message component side, most of the time there is no offset and thus a default return value of 0 (zero) is appropriate.

Fix Description:


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@23721 85c8ee13-dc6b-4ec3-b49e-93a6737f2991