BugID(s): VOIP-64473
Problem Description:
Add an attribute to [last_*] (ie [last_rseq], but not [last_cseq_method], etc keyword that allows specifying that the we want only the value and not the header itself (ie [last_rseq value_only="true"]

Fix Description:
Added the value_only attribute, and a parser to read it when the scenario is being set up. Changed the way get_last_header runs, to accept a boolean, which says whether or not to only get the value.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@23785 85c8ee13-dc6b-4ec3-b49e-93a6737f2991