BugID(s): VOIP-64473
Problem Description:
Many times 'Printintg quit message, quitting = 1' is displayed (rather than once).

Fix Description:
This message is fairly misleading as it is, since the quitting message is printed pretty much from the beginning of the call, when the last call that can be started is started (in our cases 1 call is run). With this in mind I simply removed the message entirely.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@24235 85c8ee13-dc6b-4ec3-b49e-93a6737f2991