BugID(s): VOIP-64473
Problem Description:
Add support for VLan

Fix Description:
Added a check to prepare_pcap which checks to see whether the ethernet frame is a vlan frame. If it is it checks in the vlan header to check that there is either an IP packet, or another vlan header.  No tests were done with stacked vlan.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@24647 85c8ee13-dc6b-4ec3-b49e-93a6737f2991