BugID(s): VOIP-64473
Problem Description:
No longer auto-generate remote_tag if unspecified; Implement generated="true"|"false" attribute for [remote_tag] and [remote_tag_param] such that it is auto-generated ONLY if generated="true" is specified.

Fix Description:
added a 'generated' attribute to the component, and parse it in the case when we have a remote_tag or remote_tag_param (or peer_tag, since that's the same thing). If generated is true sipp autogenerates peer tag, otherwise it doesn't.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@24967 85c8ee13-dc6b-4ec3-b49e-93a6737f2991