BugID(s): VOIP-64473
Problem Description:
Ensure SIPp return codes are sufficient to understand the reason for exit and use those reason codes to make the Rsipp output english (not just exising 'FAIL: error 255'). (And in rsipp if it couldn't launch sipp, say so nicely). So this requires complimentary changes to SIPp and to rsipp.pl.

Fix Description:
Diversified the error codes in sipp slightly to make them easier to describe.  Added said description to rsipp.pl


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@24985 85c8ee13-dc6b-4ec3-b49e-93a6737f2991