BugID(s): VOIP-64473
Problem Description:
Ensure that check_it_inverse does not trigger the 'variable used one time error' for all functions that have it.

Fix Description:
Also modified "test" so that it doesn't trigger the error either.  Added a helper function "parseCheckIt" to clean up nearly identical code in regex and test.  Takes the variable name of the variable set to assign_to, and the action the attribute belongs to. Looks for "check_it" or "check_it_inverse", if found stores the boolean value and "gets" the variable  to increment it's usage count.  Also modified the 'variable used one time error' to be more descriptive.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@25014 85c8ee13-dc6b-4ec3-b49e-93a6737f2991