BugID(s): VOIP-64473
Problem Description:
Increment unexpected message count (and display a message on-screen at least momentarily) when the -aa code path is run. (It used to increment unuexpected count, show an error, keep running but return a fail code; now we want it to increment unexpected count, show a message (not an error), keep running and not trigger a fail code). If the usage count is what triggered the failing result then maybe keep it in a new variable (ie autoresponded_message_count) and display the sum of unexpected+autoresponded in the unexpected column. Or some other aesthetic solution.

Fix Description:
Added a MESSAGE function, which simply displays a message to the screen. Support for this function was implemented in screen.cpp.
Changed "Last Error:" to "Last Message:" when a message is printed to the screen.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@25071 85c8ee13-dc6b-4ec3-b49e-93a6737f2991