BugID(s): VOIP-64473
Problem Description:
All messages written to stderr must also be written to the error log.

Fix Description:
Delayed processing of the inputted files until after the error logs are opened in order to add error messaging for these files. Also moved the definitions of the functions so they can be called by the c files.


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@25196 85c8ee13-dc6b-4ec3-b49e-93a6737f2991